+++ packages/cofd/src/dictionary/index.ts
@@ -4,3 +4,4 @@ export * from "./merits.ts";
 export * from "./anchors.ts";
 export * from "./changeling.ts";
 export * from "./werewolf.ts";
+export * from "./vampire.ts";